sample and data processing id linkage...
In order to completely disambiguate which sample and data processing was used to produce a specific clone in a Repertoire, it is necessary to have both a sample_processing_id and a data_processing_id. A Repertoire can have multiple sample processing stages performed (e.g. A repertoire is divided into a number of cell subsets) and can have multiple data processing steps applied (different sets of clones can be generated using different tools from the same repertoire).