minimal clones and trees schema for initial release (#294)
* remove custom fields and assign 'required'

* Added `repertoire_id` to `Clone`.

* Rename `Node` fields for consistence and move junction sequence fields to `Node`.

* Updates based on discussion (#302)

* Updates based on discussion

Most of the changes are based on the consensus of discussions with @eharkins and @javh. A few other minor things:
* I added `repertoire_id` and `data_processing_id` as required in `Clone`
* Do we need explicit `required: false` tags the way we seem to be doing for `miairr`?
* Is list a valid `type` for `sequence_id` in `Clone`?

* change 'list' to 'array'

* add items type to Clone.sequence_id list

* sample and data processing id linkage...

In order to completely disambiguate which sample and data processing was used to produce a specific clone in a Repertoire, it is necessary to have both a sample_processing_id and a data_processing_id. A Repertoire can have multiple sample processing stages performed (e.g. A repertoire is divided into a number of cell subsets) and can have multiple data processing steps applied (different sets of clones can be generated using different tools from the same repertoire).

* Final suggested changes.

* Put junction fields back into Clone spec and add description to various fields clarifying relationships to inferred ancestor.

Co-authored-by: Jason Vander Heiden <jason.vanderheiden@yale.edu>
Co-authored-by: Chaim Schramm <scharch@users.noreply.github.com>
Co-authored-by: Brian Corrie <bdcorrie@gmail.com>