Fixes to confusion around repository limits
The wording of this section was confusing, people are interpreting it as there is a limit to the amount of data that you can retrieve in a single query (in particular related to downloads). I don't think we want to suggest that there is a fixed limit for all repositories, just that a repository can impose a limit if it chooses to. I removed the two sections on specific sized limits and the direct Mongo discussion, as this was being interpreted that that is the limit for all repositories and moved that text to the "Data repository specific limits" section.

@schristley any objections to that. I don't think it changes the intent, but clarifies that repositories can have limits if they need them and why they might impose them.